fix algo + session

+++ start/routes.js
@@ -20,45 +20,15 @@ const User = use('App/Models/User')
 const Game = use('App/Models/Game')
 const Kill = use('App/Models/Kill')
 const Database = use('Database')
-Route.on('/').render('welcome');
+Route.get('/', 'GameController.index');
 
 /**
  * App
  */
-Route.post('game', 'GameController.store').as('game')
-Route.post('games/:id/start', 'GameController.start').as('game.start')
-Route.get('g/:encrypted', 'GameController.index')
+Route.post('/game', 'GameController.store').as('game')
+Route.post('/games/:id/start', 'GameController.start').as('game.start')
+Route.get('/g/:encrypted', 'GameController.index')
 
-/**
- * API
- */
-Route.get('api/users', async () => {
-    return await User.all();
-})
-
-Route.get('api/tokens', async () => {
-    return await Token.all()
-})
-
-Route.get('api/games', async () => {
-    console.log('Entering Games Route');
-    return await Game.query().with('kills').fetch()
-})
-
-Route.get('api/kills', async () => {
-    return await Kill.all()
-})
-
-Route.post('api/games', async ({request, auth}) => {
-    const gameData = request.only(['name'])
-    const game = await Game.create(gameData);
-    return await auth.user.games().save(game)
-})
-
-Route.get('api/games/:id', async ({params}) => {
-    const game = await Game.find(params.id);
-    return await game.toJSON();
-})
 
 Route.get('api/games/:id/start', async ({params}) => {
     console.log("Entering start");
@@ -70,20 +40,6 @@ Route.get('api/games/:id/start', async ({params}) => {
     return await game.start();
 })
 
-// Set as debug routes.
-Route.get('api/games/:id/steal', async ({session, auth, params}) => {
-    const game = await Game.find(params.id)
-
-    // Find previous owner's kill.
-    const owner = await game.owner().fetch()
-    let res = await Kill.query().where('user_id', owner.id).where('game_id', game.id).fetch()
-    
-    // Replace with new user.
-    await auth.user.kills().save(res.rows[0]);
-
-    // Change owner.
-    return await auth.user.games().save(game)
-})
 
 
 
@@ -92,7 +48,7 @@ Route.post('api/games/:id/kills', async ({request, auth, params}) => {
         // If user has already submitted a kill for this game, return.
         const existing = await Kill.query().where('user_id', auth.user.id).where('game_id', params.id).fetch();
         if (existing.rows.length)
-            return console.log('Already existing kill for user ang game id.');
+            return console.log('Already existing kill for user and game id.');
 
         // Otherwise, save kill.
         const killData = request.only(['description'])
@@ -107,18 +63,6 @@ Route.post('api/games/:id/kills', async ({request, auth, params}) => {
 })
 
 
-Route.get('api/games/:id/kills', async ({params}) => {
-    const game = await Game.find(params.id)
-
-    return await game.kills().fetch();
-})
-
-Route.get('api/games/:id/graph', async ({params}) => {
-    const game = await Game.find(params.id)
-
-    const kills = (await game.kills().fetch()).rows;
-    return kills.map(k => `${k.killer_id}:${k.victim_id}`);
-})
 
 
 Route.get('api/games/:id/kills/start', async ({params}) => {
@@ -141,8 +85,72 @@ Route.get('api/games/:id/kills/start', async ({params}) => {
 })
 
 
-Route.delete('games/:id', async ({params}) => {
-    const game = await Game.find(params.id)
-    return console.log(await game.delete())
-})
 
+
+/**
+ * API
+ */
+
+if (process.env.ENV == 'dev') {
+
+    Route.get('api/users', async () => {
+        return await User.all();
+    })
+
+    Route.get('api/tokens', async () => {
+        return await Token.all()
+    })
+
+    Route.get('api/games', async () => {
+        console.log('Entering Games Route');
+        return await Game.query().with('kills').fetch()
+    })
+
+    Route.get('api/kills', async () => {
+        return await Kill.all()
+    })
+
+    Route.post('api/games', async ({request, auth}) => {
+        const gameData = request.only(['name'])
+        const game = await Game.create(gameData);
+        return await auth.user.games().save(game)
+    })
+
+    Route.get('api/games/:id', async ({params}) => {
+        const game = await Game.find(params.id);
+        return await game.toJSON();
+    })
+    // Set as debug routes.
+    Route.get('api/games/:id/steal', async ({session, auth, params}) => {
+        const game = await Game.find(params.id)
+
+        // Find previous owner's kill.
+        const owner = await game.owner().fetch()
+        let res = await Kill.query().where('user_id', owner.id).where('game_id', game.id).fetch()
+        
+        // Replace with new user.
+        await auth.user.kills().save(res.rows[0]);
+
+        // Change owner.
+        return await auth.user.games().save(game)
+    })
+
+    Route.get('api/games/:id/kills', async ({params}) => {
+        const game = await Game.find(params.id)
+
+        return await game.kills().fetch();
+    })
+
+    Route.get('api/games/:id/graph', async ({params}) => {
+        const game = await Game.find(params.id)
+
+        const kills = (await game.kills().fetch()).rows;
+        return kills.map(k => `${k.killer_id}:${k.victim_id}`);
+    })
+
+    Route.delete('games/:id', async ({params}) => {
+        const game = await Game.find(params.id)
+        return console.log(await game.delete())
+    })
+
+}
